requirements.txt
# JEPA-DRONE Dependencies
# Data Processing
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Development
ipython>=8.0.0
jupyter>=1.0.0

# Optional: GPU acceleration
# Install separately based on your CUDA version:
# pip install torch --index-url https://download.pytorch.org/whl/cu118
